refactor: share the shader test fixtures, and make clippy pass
The GPU tests carried six copies of `word`/`block`, two each of
`passthrough_vertex_shader` and `solid_fragment_shader`, and three
transcriptions of the same solid.frag capture. They all come from
`gpu::testing` now, and `raster`'s harness delegates to
`testing::Harness` instead of repeating its register programming.
`solid_fragment_shader` carries the capture's real `bra`/`nop` padding
rather than zeros, so the test that proves decoding stops at `exit` has
padding to stop before.

`cargo clippy --workspace --all-targets` failed on two deny-level lints
and reported 262 warnings. Most of what is here are ordinary fixes; the
rest is calibration, because three of the lints are wrong for this
codebase. The opus modules are a transcription of RFC 6716, where the
tables are normative and the rounding has to match the encoder, so they
are scoped out by a macro rather than edited. Hardware encodings are
written field by field, which is what `identity_op` and
`unusual_byte_groupings` object to, so both are off workspace-wide. And
every exported function in switch-wasm takes pointers JS owns, which is
what `not_unsafe_ptr_arg_deref` sees.

One real bug: the A32 register dump printed the four GE flags as
`{:#03b}`, a width of three that the `0b` prefix takes two of, so one
flag never reached the output.

a32_neon_test::the_floating_point_arithmetic_runs_lane_by_lane already
fails at d7a6a71 and is untouched here.

+++ crates/switch-core/src/disasm.rs
@@ -671,9 +671,9 @@ fn disasm_dp_imm(insn: u32, s: &mut String) -> Result<bool, std::fmt::Error> {
             if ((insn >> 23) & 1) == 0 {
                 let opc = (insn >> 29) & 0b11;
                 let (immr, imms) = if sf == 1 {
-                    (((insn >> 16) & 0x3F) as u32, ((insn >> 10) & 0x3F) as u32)
+                    ((insn >> 16) & 0x3F, (insn >> 10) & 0x3F)
                 } else {
-                    (((insn >> 16) & 0x1F) as u32, ((insn >> 10) & 0x1F) as u32)
+                    ((insn >> 16) & 0x1F, (insn >> 10) & 0x1F)
                 };
                 let (lsb, msb) = (immr as i64, imms as i64);
                 let datasize = if sf == 1 { 64 } else { 32 };
